+++ Classes/Connection/MatomoConnector.php
@@ -21,11 +21,11 @@ use TYPO3\CMS\Core\Http\Stream;
 /**
  * @internal
  */
-class MatomoConnector
+readonly class MatomoConnector
 {
     public function __construct(
-        private readonly RequestFactoryInterface $requestFactory,
-        private readonly GuzzleClientFactory $guzzleClientFactory,
+        private RequestFactoryInterface $requestFactory,
+        private GuzzleClientFactory $guzzleClientFactory,
     ) {}
 
     public function callApi(ConnectionConfiguration $configuration, string $method, ParameterBag $parameterBag): array
@@ -34,7 +34,6 @@ class MatomoConnector
             ->set('module', 'API')
             ->set('idSite', (string) $configuration->idSite)
             ->set('method', $method)
-            ->set('token_auth', $configuration->tokenAuth)
             ->set('format', 'json');
 
         $body = new Stream('php://temp', 'r+');
@@ -42,6 +41,7 @@ class MatomoConnector
 
         $request = $this->requestFactory->createRequest('POST', $configuration->url)
             ->withHeader('content-type', 'application/x-www-form-urlencoded')
+            ->withHeader('authorization', 'Bearer ' . $configuration->tokenAuth)
             ->withBody($body);
 
         $response = $this->guzzleClientFactory->getClient()->send($request);