+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Matomo 5.4+ is required (#60)
+
 ### Removed
 - Compatibility with TYPO3 v12 (#53)